storage: Move storage classes to subdir for cleaner separation
