+++ exactonline/api/v1division.py
@@ -9,7 +9,7 @@ Copyright (C) 2015-2017 Walter Doekes, OSSO B.V.
 from ..exceptions import ExactOnlineError
 from ..http import urljoin
 from ..resource import GET
-from ..storage import NoOptionError
+from ..storage import MissingSetting
 
 
 class V1DivisionError(ExactOnlineError):
@@ -20,7 +20,7 @@ class V1Division(object):
     def restv1(self, request):
         try:
             division = self.storage.get_division()
-        except NoOptionError:
+        except MissingSetting:
             raise V1DivisionError('Division unset/blank in config')
         if not division:
             raise V1DivisionError('Division unset/blank in config')
