+++ exactonline/storage_test.py
@@ -1,53 +1,16 @@
 # vim: set ts=8 sw=4 sts=4 et ai tw=79:
 """
-Provides a storage class to the Exact Online REST API Library.
+Test ini storage class to the Exact Online REST API Library.
 
 This file is part of the Exact Online REST API Library in Python
 (EORALP), licensed under the LGPLv3+.
-Copyright (C) 2015-2017 Walter Doekes, OSSO B.V.
-
-Usage:
-
-    storage = IniStorage('somefile.ini')
-
-Or, if you want to use your own storage backend, you may provide a set()
-and get() method and inherit from the ExactOnlineConfig.
-
-    class MyStorage(ExactOnlineConfig):
-        def get(self, section, option):
-            try:
-                return self.fetch_stuff_from_somewhere(...)
-            except:
-                raise NoOptionError()
-
-        def set(self, section, option, value):
-            self.store_stuff_somewhere(...)
-
-Example ini file:
-
-    [server]
-    auth_url = https://start.exactonline.co.uk/api/oauth2/auth
-    rest_url = https://start.exactonline.co.uk/api
-    token_url = https://start.exactonline.co.uk/api/oauth2/token
-
-    [application]
-    base_url = https://example.com
-    client_id = {12345678-abcd-1234-abcd-0123456789ab}
-    client_secret = ZZZ999xxx000
-
-    [transient]
-    access_expiry = 1426492503
-    access_token = dAfjGhB1k2tE2dkG12sd1Ff1A1fj2fH2Y1j1fKJl2f1sD1ON275zJNUy...
-    code = dAfj!hB1k2tE2dkG12sd1Ff1A1fj2fH2Y1j1fKJl2f1sD1ON275zJNUy...
-    division = 123456
-    refresh_token = SDFu!12SAah-un-56su-1fj2fH2Y1j1fKJl2f1sDfKJl2f1sD11FfUn1...
-
+Copyright (C) 2015-2018 Walter Doekes, OSSO B.V.
 """
 from unittest import TestCase, main
 from io import StringIO
 from os import path, unlink
 
-from .storage import IniStorage, NoOptionError
+from .storage.ini import IniStorage, MissingSetting
 
 
 class IniStorageTestCase(TestCase):
@@ -64,7 +27,7 @@ class IniStorageTestCase(TestCase):
 
     def test_missing_section_raises_nooption(self):
         config = IniStorage(StringIO())
-        self.assertRaises(NoOptionError, config.get, 'bad_section',
+        self.assertRaises(MissingSetting, config.get, 'bad_section',
                           'bad_value')
 
     def test_server_defaults(self):
@@ -96,18 +59,18 @@ class IniStorageTestCase(TestCase):
 
     def test_application_no_defaults(self):
         config = IniStorage(StringIO())
-        self.assertRaises(NoOptionError, config.get_base_url)
-        self.assertRaises(NoOptionError, config.get_response_url)  # same..
-        self.assertRaises(NoOptionError, config.get_client_id)
-        self.assertRaises(NoOptionError, config.get_client_secret)
+        self.assertRaises(MissingSetting, config.get_base_url)
+        self.assertRaises(MissingSetting, config.get_response_url)  # same..
+        self.assertRaises(MissingSetting, config.get_client_id)
+        self.assertRaises(MissingSetting, config.get_client_secret)
 
     def test_transient_no_defaults(self):
         config = IniStorage(StringIO())
-        self.assertRaises(NoOptionError, config.get_access_expiry)
-        self.assertRaises(NoOptionError, config.get_access_token)
-        self.assertRaises(NoOptionError, config.get_code)
-        self.assertRaises(NoOptionError, config.get_division)
-        self.assertRaises(NoOptionError, config.get_refresh_token)
+        self.assertRaises(MissingSetting, config.get_access_expiry)
+        self.assertRaises(MissingSetting, config.get_access_token)
+        self.assertRaises(MissingSetting, config.get_code)
+        self.assertRaises(MissingSetting, config.get_division)
+        self.assertRaises(MissingSetting, config.get_refresh_token)
 
     def test_example_ini(self):
         config = IniStorage(path.join(path.dirname(__file__), 'example.ini'))
@@ -145,7 +108,7 @@ class IniStorageTestCase(TestCase):
         try:
             self.assertFalse(path.exists('storage-gets-created.ini'))
             config = IniStorage('storage-gets-created.ini')
-            self.assertRaises(NoOptionError, config.get_division)
+            self.assertRaises(MissingSetting, config.get_division)
             config.set_division(987654321)
             self.assertEqual(config.get_division(), 987654321)
             config.set_division(11223344)
